fix(harm-floor): annotate monthly_cost as int (review follow-up)
Money-change review found the only real (if minor) issue: Component.monthly_cost
(-> float) and SimulationResult.monthly_cost (: float) declared float while now
provably always int (both factors are validated ints; cost lines are pure-integer).
Annotate them int so the type contract matches ADR-008. Harm-floor + no-drift lenses
came back CLEAN; no runtime change. Gate: 253 tests green.

+++ prototype/keystone/model.py
@@ -102,7 +102,8 @@ class Component:
         return self.per_instance_rps * self.instances
 
     @property
-    def monthly_cost(self) -> float:
+    def monthly_cost(self) -> int:
+        # int × int → int: both factors are validated ints, so money stays integer cents (harm floor).
         return self.monthly_cost_per_instance * self.instances
 
     @property
